+++ editor/editor-plugin-layout/src/pm-plugins/actions.ts
@@ -441,7 +441,12 @@ export const setPresetLayout =
 	(layout: PresetLayout): Command =>
 	(state, dispatch) => {
 		const { pos, selectedLayout } = pluginKey.getState(state) as LayoutState;
-		if (selectedLayout === layout || pos === null) {
+
+		if (fg('platform_editor_advanced_layouts_post_fix_patch_3')) {
+			if (pos === null) {
+				return false;
+			}
+		} else if (selectedLayout === layout || pos === null) {
 			return false;
 		}
 
@@ -450,7 +455,13 @@ export const setPresetLayout =
 			return false;
 		}
 
-		let tr = forceSectionToPresetLayout(state, node, pos, layout);
+		if (fg('platform_editor_advanced_layouts_post_fix_patch_3')) {
+			if (selectedLayout === layout && node.childCount > 1) {
+				return false;
+			}
+		}
+
+		const tr = forceSectionToPresetLayout(state, node, pos, layout);
 		if (tr) {
 			editorAnalyticsAPI?.attachAnalyticsEvent({
 				action: ACTION.CHANGED_LAYOUT,
@@ -485,6 +496,7 @@ const getDefaultPresetLayout = (layoutNode: Node): PresetLayout => {
 	if (layoutColumnCount <= 1) {
 		// This prevents the creation of a single column layout
 		// once we support single column layout, we can return 'single'
+
 		return fg('platform_editor_advanced_layouts_post_fix_patch_1') ? 'two_equal' : 'single';
 	}
 	switch (layoutColumnCount) {
@@ -587,7 +599,7 @@ export const deleteActiveLayoutNode =
 		if (pos !== null) {
 			const node = state.doc.nodeAt(pos) as Node;
 			if (dispatch) {
-				let tr = state.tr.delete(pos, pos + node.nodeSize);
+				const tr = state.tr.delete(pos, pos + node.nodeSize);
 				editorAnalyticsAPI?.attachAnalyticsEvent({
 					action: ACTION.DELETED,
 					actionSubject: ACTION_SUBJECT.LAYOUT,