+++ editor/collab-provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/collab-provider
 
+## 10.3.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.3.1
 
 ### Patch Changes