+++ editor/editor-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-core
 
+## 203.8.0
+
+### Minor Changes
+
+- [#98971](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/pull-requests/98971)
+  [`1af472c6db66f`](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/commits/1af472c6db66f) -
+  Remove platform_editor_context_panel_support_multiple fg
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 203.7.5
 
 ### Patch Changes