+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @atlaskit/editor-plugins
 
+## 6.1.1
+
+### Patch Changes
+
+- Updated dependencies
+
+## 6.1.0
+
+### Minor Changes
+
+- [#98971](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/pull-requests/98971)
+  [`1af472c6db66f`](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/commits/1af472c6db66f) -
+  Remove platform_editor_context_panel_support_multiple fg
+
+### Patch Changes
+
+- Updated dependencies
+
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 6.0.1
 
 ### Patch Changes